+++ src/pages/Acknowledgments/index.ts
@@ -1,0 +1,1 @@
+export { Acknowledgments } from "./Acknowledgments";